Added ActiveRecord->hasErrors Inverted parameter order on html->link and html->button

+++ branches/reboot/php2go/activeRecord/relation/ActiveRecordRelationBelongsTo.php
@@ -3,8 +3,8 @@
 class ActiveRecordRelationBelongsTo extends ActiveRecordRelation
 {
 	private static $cache = array();
-	protected $requiredOptions = array('class', 'foreignKey');	
-	
+	protected $requiredOptions = array('class', 'foreignKey');
+
 	public function merge(ActiveRecord $base, array &$criteria) {
 		$this->loaded = true;
 		$db = Db::instance();
@@ -14,7 +14,7 @@ class ActiveRecordRelationBelongsTo extends ActiveRecordRelation
 		ActiveRecord::normalizeAliases($model, $criteria, $this->name);
 		$criteria['join'][] = $this->buildJoin($base, $model);
 	}
-	
+
 	public function find(ActiveRecord $base) {
 		$model = ActiveRecord::model($this->options['class']);
 		$key = $model->getTableName() . '-' . $base->{$this->options['foreignKey']};
@@ -22,9 +22,9 @@ class ActiveRecordRelationBelongsTo extends ActiveRecordRelation
 			self::$cache[$key] = DAO::instance()->find($model->getTableName(), sprintf('%s = ?', $model->getMetaData()->primaryKey), array($base->{$this->options['foreignKey']}));
 		return self::$cache[$key];
 	}
-	
+
 	public function save(ActiveRecord $base, $model) {
-		if ($model instanceof ActiveRecord) {			
+		if ($model instanceof ActiveRecord) {
 			if ($model->save()) {
 				$base->{$this->options['foreignKey']} = $model->getPrimaryKey();
 				return true;
@@ -35,22 +35,22 @@ class ActiveRecordRelationBelongsTo extends ActiveRecordRelation
 		}
 		return true;
 	}
-	
+
 	public function delete(ActiveRecord $base) {
-		if ($this->options['deleteCascade'] === true) {
+		if (isset($this->options['deleteCascade']) && $this->options['deleteCascade'] === true) {
 			$model = $base->getRelation($this->name);
 			if ($model instanceof ActiveRecord)
 				return $model->delete();
 		}
 		return true;
 	}
-	
+
 	protected function buildJoin(ActiveRecord $base, ActiveRecord $model) {
 		return sprintf(
-			"inner join %s on %s.%s = %s.%s", 
+			"inner join %s on %s.%s = %s.%s",
 			($this->name != $model->getTableName() ? "{$model->getTableName()} as {$this->name}" : $this->name),
 			$base->getTableName(), $this->options['foreignKey'],
 			$this->name, $model->getMetaData()->primaryKey
 		);
-	}	
+	}
 }
